feat(tax): widen tax_rate_bp to BIGINT + align Go types (HYG-3)
Closes out the tax-rate representation item. The functional half of
HYG-3 already shipped in migration 0003, which dropped the legacy
NUMERIC(6,2) / NUMERIC(5,4) tax_rate columns so all tax arithmetic now
runs through integer basis points end-to-end — no floats anywhere in
the money path.

Remaining gap: tax_rate_bp was INT4 while every peer money column
(tax_amount_cents, subtotal_cents, total_amount_cents, …) is BIGINT.
INT4's 21-million-percent ceiling is not a real overflow risk, but the
schema inconsistency is exactly the "why is this one column different?"
trap the HYG- items exist to remove.

Migration 0014 widens tax_rate_bp on tenant_settings, invoices, and
invoice_line_items with ALTER … TYPE BIGINT (metadata-only on Postgres
12+, safe at any table size). Go-side TaxRateBP / taxRateBP fields flip
from int to int64 and TaxOverrideRateBP from *int to *int64 so the
domain types match the column width exactly.

+++ internal/billing/engine_tax_apply_test.go
@@ -14,7 +14,7 @@ import (
 // NextInvoiceNumber() is unused by ApplyTaxToLineItems so it errors to catch
 // any accidental call.
 type taxSettings struct {
-	rateBP    int
+	rateBP    int64
 	name      string
 	inclusive bool
 }
@@ -52,7 +52,7 @@ func (c *stubCalculator) CalculateTax(_ context.Context, _ string, _ tax.Custome
 	return c.result, c.err
 }
 
-func newTaxTestEngine(rateBP int, name string, profiles map[string]domain.CustomerBillingProfile) *Engine {
+func newTaxTestEngine(rateBP int64, name string, profiles map[string]domain.CustomerBillingProfile) *Engine {
 	e := &Engine{
 		settings: &taxSettings{rateBP: rateBP, name: name},
 	}
@@ -62,7 +62,7 @@ func newTaxTestEngine(rateBP int, name string, profiles map[string]domain.Custom
 	return e
 }
 
-func newInclusiveTaxTestEngine(rateBP int, name string) *Engine {
+func newInclusiveTaxTestEngine(rateBP int64, name string) *Engine {
 	return &Engine{settings: &taxSettings{rateBP: rateBP, name: name, inclusive: true}}
 }
 
@@ -89,7 +89,7 @@ func TestApplyTaxToLineItems_TenantRate(t *testing.T) {
 }
 
 func TestApplyTaxToLineItems_CustomerOverrideWins(t *testing.T) {
-	override := 2500 // 25%
+	override := int64(2500) // 25%
 	profiles := map[string]domain.CustomerBillingProfile{
 		"cus_1": {
 			CustomerID:        "cus_1",
